7. Cloud Computing for Data Engineer Projects/Loan Fraud Monitoring/requirements.txt
faker
azure-eventhub